location-register/app/requirements.txt
pyjwt[crypto]
python-keycloak
python-multipart